feat(assertion): refactoring generated folder assertion setup - part 2
